% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_expectation_k.R
\name{approximate_expectation_k}
\alias{approximate_expectation_k}
\title{Approximate expectation of k given number of clinical samples and replicates}
\usage{
approximate_expectation_k(n, R)
}
\arguments{
\item{n}{Integer - The number of unique clinical samples}

\item{R}{Integer - The number of replicated measurements done on each clinical sample}
}
\value{
A value that approximates the true expectation of k
}
\description{
Approximate expectation of k given number of clinical samples and replicates
}
\examples{
approximate_expectation_k(25,3)
}
