% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approximate_variance_k.R
\name{approximate_variance_k}
\alias{approximate_variance_k}
\title{Approximate variance of k given number of clinical samples and replicates}
\usage{
approximate_variance_k(n, R)
}
\arguments{
\item{n}{Integer - The number of unique clinical samples}

\item{R}{Integer - The number of replicated measurements done on each clinical sample}
}
\value{
A value that approximates the true variance of k
}
\description{
Approximate variance of k given number of clinical samples and replicates
}
\examples{
approximate_variance_k(25, 3)
}
