% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_k_values_over_groups.R
\name{attach_k_values_over_groups}
\alias{attach_k_values_over_groups}
\title{Attach calculated k values for each group}
\usage{
attach_k_values_over_groups(data, k_data, groups = "Comparison")
}
\arguments{
\item{data}{A grouped data frame or data table with format LFDT that is the one we want the k values attached to}

\item{k_data}{A grouped (same grouping variables as data) data frame or data table with format LFDT containing the k values}

\item{groups}{The names of the grouping columns of data and k_data}
}
\value{

}
\description{
Attach calculated k values for each group
}
\examples{
attach_k_values_over_groups(data = Deming_CLSI_EP14_over_groups(MS_wise(sampled_cs_measurements), groups = "Comparison", level = 0.95, Np = 1, evaluated_materials = MS_wise(sampled_eqam_measurements)), k_data = estimate_k_over_groups(MS_wise(sampled_cs_measurements)))
}
