% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deming_CLSI_EP14_over_groups.R
\name{Deming_CLSI_EP14_over_groups}
\alias{Deming_CLSI_EP14_over_groups}
\title{Estimate prediction band using Deming regression formulated as in EP14 over groups}
\usage{
Deming_CLSI_EP14_over_groups(
  data,
  groups = "Comparison",
  level = 0.99,
  R = 3,
  Np = 1000,
  evaluated_materials = NULL,
  column_order = c("Comparison", "SampleID", "MP_B", "MP_A", "fit", "lwr", "upr")
)
}
\arguments{
\item{data}{A data frame or data table with format LFDT enclosed with all replicated measurements}

\item{groups}{The names of the grouping columns of the data}

\item{level}{A numeric value that captures the overall confidence level of the estimated prediction band}

\item{R}{An integer signifying the maximum number of replicated measurements performed on each evaluated material}

\item{Np}{An integer, which captures the number of pointwise prediction intervals making the prediction band across the concentration range}

\item{evaluated_materials}{A data frame or data table with format LFDT enclosed with all replicated measurements for evaluated materials such as EQAMs or CRMs. Should be NULL if the PB, that is, pointwise prediction intervals to be estimated}

\item{column_order}{A vector specifying the order of the columns of the outputted grouped data table. Will be ignored if evaluated_materials is NULL. Note that the length of column_order must be the same as the number of columns of output columns, which is 6 + the number of grouping columns}
}
\value{
A grouped data table enclosed with information regarding the estimated prediction band across the concentration range or for the particular evaluated materials for all groups
}
\description{
Estimate prediction band using Deming regression formulated as in EP14 over groups
}
\details{
This is the prediction interval estimation procedure as described by CLSI in EP14. This is recommended to use over Deming_Gillard_Fuller_over_groups if the calculated k value is smaller than 1
}
\examples{
Deming_CLSI_EP14_over_groups(MS_wise(sampled_cs_measurements), groups = "Comparison", level = 0.95, Np = 1, evaluated_materials = MS_wise(sampled_eqam_measurements))
}
