% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deming_Gillard_Fuller.R
\name{Deming_Gillard_Fuller}
\alias{Deming_Gillard_Fuller}
\title{Estimate prediction band using Deming regression formulated by J. Gillard and Fuller}
\usage{
Deming_Gillard_Fuller(
  data,
  level = 0.99,
  R = 3,
  Np = 1000,
  evaluated_materials = NULL
)
}
\arguments{
\item{data}{A data frame or data table with format LFDT enclosed with all replicated measurements}

\item{level}{A numeric value that captures the overall confidence level of the estimated prediction band. This is assumed to already be Bonferroni-corrected. The recommended base level is 0.99}

\item{R}{An integer signifying the maximum number of replicated measurements performed on each evaluated material. This will shrink the width of the prediction interval and band because using the mean of replicates decreases uncertainty}

\item{Np}{An integer, which captures the number of pointwise prediction intervals making the prediction band across the concentration range. Ignored if evaluated_materials is !NULL}

\item{evaluated_materials}{A data frame or data table with format LFDT enclosed with all replicated measurements for evaluated materials such as EQAMs or CRMs. Should be NULL if the PB, that is, pointwise prediction intervals to be estimated}
}
\value{
A data table enclosed with information regarding the estimated prediction band across the concentration range or for the particular evaluated materials
}
\description{
Estimate prediction band using Deming regression formulated by J. Gillard and Fuller
}
\examples{
Deming_Gillard_Fuller(data = MS_wise(sampled_cs_measurements)[Comparison=="MP1 - MP2",], level = 0.95, R = 3, evaluated_materials = NULL)
}
