% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_k.R
\name{estimate_k}
\alias{estimate_k}
\title{Estimate a k value based on one MS comparison}
\usage{
estimate_k(data)
}
\arguments{
\item{data}{A data frame or data table with format LFDT where all replicate measurements are enclosed}
}
\value{
A data table containing a plug-in point estimate of K, namely, k
}
\description{
Estimate a k value based on one MS comparison
}
\examples{
estimate_k(data = MS_wise(sampled_cs_measurements)[Comparison=="MP1 - MP2"])
}
