% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_k_over_groups.R
\name{estimate_k_over_groups}
\alias{estimate_k_over_groups}
\title{Calculate k values over multiple groups}
\usage{
estimate_k_over_groups(data, groups = "Comparison")
}
\arguments{
\item{data}{A grouped data frame or data table with format LFDT}

\item{groups}{The names of the grouping columns of the data}
}
\value{
A data table listing the calculated k values for each group
}
\description{
Calculate k values over multiple groups
}
\examples{
estimate_k_over_groups(data = MS_wise(sampled_cs_measurements), groups = "Comparison")
}
