% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_column_names.R
\name{fix_column_names}
\alias{fix_column_names}
\title{Automatically fix column names using the first algorithm}
\usage{
fix_column_names(data, LFDT = FALSE, MOR = FALSE, ordered = TRUE, n_ids = 3)
}
\arguments{
\item{data}{A data frame or data table with format LFDT or WFDT having \code{n_ids} ID columns}

\item{LFDT}{Is the data on long-format with comparison, i.e., TRUE or on wide format without comparison which corresponds to FALSE}

\item{MOR}{Is the data only consisting of mean of replicates? If so, set this to TRUE. Or else, set this to FALSE}

\item{ordered}{Is the data ordered with \code{Comparison} first, then \code{SampleID}, and last \code{ReplicateID}? If not, data is sent to \code{fix_column_names_2()}, which does approximate matching with slack = 3. Note that setting to ordered when not being ordered in the correct way will affect the output data in some negative manner}

\item{n_ids}{The number of ID columns found in \code{data}. Ignored if \code{ordered} is \code{TRUE}}
}
\value{
A data table identical to input \code{data} but ID column names are replaced with standard ones. If ordered is \code{FALSE} then approximate matching is done with slack = 3
}
\description{
Automatically fix column names using the first algorithm
}
\examples{
fix_column_names(data = sampled_cs_measurements, ordered = TRUE)
}
