% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_column_names_2.R
\name{fix_column_names_2}
\alias{fix_column_names_2}
\title{Automatically fix column names using the second algorithm}
\usage{
fix_column_names_2(
  data,
  n_ids = 3,
  LFDT = FALSE,
  MOR = FALSE,
  slack = 3,
  print_matches = FALSE
)
}
\arguments{
\item{data}{A data frame or data table with format LFDT or WFDT having \code{n_ids} ID columns. The ID columns are not required to be ordered as is the case for \code{fix_column_names()}}

\item{n_ids}{The number of ID columns found in \code{data}}

\item{LFDT}{Is the data on long-format with comparison, i.e., TRUE or on wide format without comparison which corresponds to FALSE}

\item{MOR}{Is the data only consisting of mean of replicates? If so, set this to TRUE. Or else, set this to FALSE}

\item{slack}{How much deviance from look-up table should be allow.}

\item{print_matches}{Should the matches for the respective ID columns from \code{data} be displayed? If there are more than four matches, only the first four will be presented and the rest omitted}
}
\value{
A data table identical to input \code{data} but ID column names are potentially replaced with standard ones if they are found in respective look-up tables
}
\description{
In order to use many of the function of this package, it is required that the ID columns are given by particular names. Comparison ID column should be named \code{Comparison}, Clinical sample ID should be named \code{SampleID} and Replicate measurement ID should be named \code{ReplicateID}. If given names deviates from the standard names, one might experience problems when using some of the functions found in this package
}
\details{
You need to be careful with how you specify LFDT, MOR and n_ids as wrong combinations of these may not be supported. Contact Author for further information
}
\examples{
fix_column_names_2(sampled_eqam_measurements, LFDT = FALSE, MOR = FALSE, slack = 1, print_matches = TRUE)
}
