% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_reverse_issues.R
\name{fix_reverse_issues}
\alias{fix_reverse_issues}
\title{Based on lambda, which set of measurement system measurements should be used as x and y}
\usage{
fix_reverse_issues(data, groups = "Comparison")
}
\arguments{
\item{data}{A data table found in \code{Ordinary_Least_Squares_over_groups()}}

\item{groups}{A vector of grouping columns}
}
\value{
A fixed data table
}
\description{
Based on lambda, which set of measurement system measurements should be used as x and y
}
\examples{
fix_reverse_issues(data.table(Comparison = "f - g", SampleID = "ID 1", MP_B = 5.10, MP_A = 5.20, fit = 5.08, lwr = 5.00, upr = 5.16, reversed = TRUE))
}
