% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_optimal_binwidth.R
\name{get_optimal_binwidth}
\alias{get_optimal_binwidth}
\title{Calculate the optimal binwidths for histograms based on several methods}
\usage{
get_optimal_binwidth(data, method = c("FDR", "SNRR", "SF", "RR", "DF"))
}
\arguments{
\item{data}{A numeric vector that is used to construct histograms}

\item{method}{Which method should be used to get the optimal binwidths. Valid inputs are \code{FDR}, \code{SNRR}, \code{SF}, \code{RR} and \code{DF}. See details for more information regarding the different approaches}
}
\value{
A value that represents the optimal binwidth based on data and chosen method
}
\description{
Calculate the optimal binwidths for histograms based on several methods
}
\examples{
get_optimal_binwidth(rnorm(100), method = "SF")
}
