% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_bootstrap.R
\name{gro_bootstrap}
\alias{gro_bootstrap}
\title{bootstrap estimates}
\usage{
gro_bootstrap(
  data,
  filter_var = "NO",
  B = 1000,
  alpha = 0.05,
  statistic = median,
  calculate = c("confidence interval", "variance", "standard deviation",
    "coefficient of variation")
)
}
\arguments{
\item{data}{some data}

\item{filter_var}{some value referring to missing values}

\item{B}{how many bootstrap replicates should be used}

\item{alpha}{significance level}

\item{statistic}{Which statistic should we do inference on. A function that exists in R taking a numeric vector. median is default}

\item{calculate}{What do we want to calculate for the statistic considered. EIther \code{confidence interval}, \code{variance}, \code{standard deviation} or \code{coefficient of variation}}
}
\value{
values for each considered group
}
\description{
bootstrap estimates
}
\examples{
print(1)
}
