% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_resample.R
\name{gro_resample}
\alias{gro_resample}
\title{resample values in one specific group}
\usage{
gro_resample(data, B = 1000, fun = median)
}
\arguments{
\item{data}{a subset of the total data}

\item{B}{number of bootstrap replicates}

\item{fun}{which statistic}
}
\value{
return something cool
}
\description{
resample values in one specific group
}
\examples{
print(1)
}
