% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gro_standard.R
\name{gro_standard}
\alias{gro_standard}
\title{XXX}
\usage{
gro_standard(data, filter_var = "NO", alpha = 0.05)
}
\arguments{
\item{data}{data from GRO}

\item{filter_var}{which value represents missing value in data}

\item{alpha}{significance level for two-sided asymptotic z-interval}
}
\value{
returns something useful
}
\description{
XXX
}
\examples{
print(1)
}
