% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heteroscedasticity_test.R
\name{heteroscedasticity_test}
\alias{heteroscedasticity_test}
\title{Testing for variance-inhomogeneity using Breauch-Pagan test for each comparison}
\usage{
heteroscedasticity_test(
  data,
  alpha = 0.05,
  groups = "Comparison",
  bonferroni_correct = FALSE,
  conservative = TRUE,
  robust = TRUE
)
}
\arguments{
\item{data}{A data table or data frame of type LFDT grouped according to \code{groups}}

\item{alpha}{The base significance level of the test.}

\item{groups}{The names of grouping columns in \code{data}}

\item{bonferroni_correct}{Should we Bonferroni correct the significance level based the number of MS comparisons. Default is \code{TRUE}}

\item{conservative}{Should the Bonferroni correction be done conservatively that assumes that some of the MS comparisons are dependent? Default is \code{TRUE}}

\item{robust}{Should Koenker's adjust be implemented in the test? Recommended if data is non-normal. Default is \code{TRUE}}
}
\value{
A data table with grouping as the first column, LM test statistic as the second column, p-value as the third column, and judgement as the last column
}
\description{
Testing for variance-inhomogeneity using Breauch-Pagan test for each comparison
}
\examples{
print(1)
}
