% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sdwdnsp1.R
\name{mc_sdwdnsp1}
\alias{mc_sdwdnsp1}
\title{Simulate several data sets with different non-selectivity profiles defined by setting 1 using parallel computation}
\usage{
mc_sdwdnsp1(
  parameter_row = c(1),
  m = 100,
  n = "r",
  R = "r",
  cv_vec = "r",
  ci = "r",
  mmax = 2.5,
  p = 0.05,
  parallel = TRUE
)
}
\arguments{
\item{parameter_row}{A data frame or data table with only 1 row. Should contain a subset of simulation parameters. Parameters that may be included are n, R, cvx, cvy, ci_lwr, ci_upr, mmax and p}

\item{m}{Integer - The number of data sets to be simulated}

\item{n}{Integer or 'r' - number of unique samples}

\item{R}{Integer or 'r' - Number of unique replicated measurements on each sample}

\item{cv_vec}{Vector with two elements or 'r' - MS CVs in decimal}

\item{ci}{Vector with two elements or 'r' - Concentration interval}

\item{mmax}{A number signifying the maximum multiple of relocation of clinical samples affected by differences in selectivity}

\item{p}{A number signifying the average proportion of clinical samples to be relocated}

\item{parallel}{allow parallelization}
}
\value{
A list of simulated data tables with 4 columns, where two of which are measurement results from two MSs in comparison having differences in selectivity profiles. In addition, n * R rows
}
\description{
Simulate several data sets with different non-selectivity profiles defined by setting 1 using parallel computation
}
\details{
this function is recommended to use when simulating more than 350 data sets, or else it will be faster to just use standard \code{sapply()}
}
\examples{
mc_sdwdnsp1()
}
