% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sdwinsp.R
\name{mc_sdwinsp}
\alias{mc_sdwinsp}
\title{Simulate several data sets with identical non-selectivity profiles using parallel computation}
\usage{
mc_sdwinsp(
  parameter_row = c(2),
  m = 100,
  n = "r",
  R = "r",
  cv_vec = "r",
  ci = "r",
  parallel = FALSE,
  progress_bar = TRUE
)
}
\arguments{
\item{parameter_row}{A data frame or data table with only 1 row. Should contain a subset of simulation parameters. Possible simulation parameters are n, R, cvx, cvy, ci_lwr and ci_upr}

\item{m}{Integer - The number of data sets to be simulated}

\item{n}{Integer or 'r' - number of unique samples}

\item{R}{Integer or 'r' - Number of unique replicated measurements on each sample}

\item{cv_vec}{Vector with two elements or 'r' - MS CVs in decimal}

\item{ci}{Vector with two elements or 'r' - Concentration interval}

\item{parallel}{Should we allow parallel computing to be performed when simulating from \code{sdwinsp()}. This is not recommended if the number of parameter combinations is larger than 10. Default value is FALSE}

\item{progress_bar}{Should a progress bar monitoring the simulation progress be displayed. Default is TRUE}
}
\value{
A list containing m data tables with 4 columns, two of which are measured results from two MSs in comparison. In addition, n * R rows.
}
\description{
Simulate several data sets with identical non-selectivity profiles using parallel computation
}
\details{
this function is recommended to use when simulating more than 1000 data sets, or else it will be faster to just use standard sapply
}
\examples{
mc_sdwinsp(m = 50)
}
