% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_replicates.R
\name{mean_of_replicates}
\alias{mean_of_replicates}
\title{Calculate mean of replicates based on grouping}
\usage{
mean_of_replicates(data, IDs = c("Comparison", "SampleID"))
}
\arguments{
\item{data}{A data frame or data table with format LFDT where all replicated measurements are enclosed}

\item{IDs}{The names of the grouping columns of the data. The column of replicate IDs must not be part of these names}
}
\value{
A data table with format LFDT where replicated measurements are collapsed to mean of replicated measurements
}
\description{
Calculate mean of replicates based on grouping
}
\examples{
mean_of_replicates(data = MS_wise(sampled_eqam_measurements)[Comparison=="MP1 - MP2",])
}
