% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_of_replicates_2.R
\name{mean_of_replicates_2}
\alias{mean_of_replicates_2}
\title{Calculate mean of replicates over Sample IDs and possibly other grouping columns specified}
\usage{
mean_of_replicates_2(data, other_ids = NULL)
}
\arguments{
\item{data}{A data frame or data table with format WFDT where all replicated measurements are enclosed}

\item{other_ids}{If additional grouping columns should be included, you include them by name here. If your SampleID and ReplicateID columns are named something else, then you need to specify their actual names here}
}
\value{
A data table with format WFDT where replicated measurements are collapsed to mean of replicated measurements
}
\description{
Calculate mean of replicates over Sample IDs and possibly other grouping columns specified
}
\examples{
mean_of_replicates_2(sampled_cs_measurements)
}
