% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_precision.R
\name{MS_precision}
\alias{MS_precision}
\title{Estimate precision of measurement systems in comparison based on replicated measurements for LFDT data}
\usage{
MS_precision(data)
}
\arguments{
\item{data}{A data frame or data table with shape LFDT where all replicated measurements are enclosed}
}
\value{
A data table with format LFDT representing precision estimates for each MS comparison. MS variances, MS CVs and lambda is returned.
}
\description{
Estimate precision of measurement systems in comparison based on replicated measurements for LFDT data
}
\examples{
MS_precision(MS_wise(sampled_cs_measurements)[Comparison=="MP1 - MP2",])
}
