% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_precision_over_groups.R
\name{MS_precision_over_groups}
\alias{MS_precision_over_groups}
\title{Estimate MS precision over multiple groups}
\usage{
MS_precision_over_groups(data, groups = "Comparison")
}
\arguments{
\item{data}{A grouped data frame or data table with format LFDT}

\item{groups}{The names of the grouping columns of the data}
}
\value{
A data table listing the estimated MS precision for each group
}
\description{
Estimate MS precision over multiple groups
}
\examples{
MS_precision_over_groups(data = MS_wise(sampled_cs_measurements))
}
