% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_wise.R
\name{MS_wise}
\alias{MS_wise}
\title{Converting standard wide format MS comparison data (WFDT) to standard long format (LFDT)}
\usage{
MS_wise(data, IDs = c("SampleID", "ReplicateID"))
}
\arguments{
\item{data}{A data frame or data table with m + n columns where m (usually two) of them are ID columns and n of them is numeric columns consisting of measurement results for each measurement system}

\item{IDs}{A character vector of length m containing the ID column names}
}
\value{
A data table of format LFDT, with m + 3 columns where the latter three columns are MS comparison ID column (new) and resulting numeric columns in this comparison
}
\description{
Converting standard wide format MS comparison data (WFDT) to standard long format (LFDT)
}
\examples{
MS_wise(sampled_eqam_measurements, IDs = c("SampleID","ReplicateID"))
}
