% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_test.R
\name{normality_test}
\alias{normality_test}
\title{Checks normality of error terms by using Shapiro Wilk's test on each groups' residuals}
\usage{
normality_test(
  data,
  alpha = 0.05,
  groups = "Comparison",
  bonferroni_correct = FALSE,
  conservative = TRUE
)
}
\arguments{
\item{data}{A data table or data frame of type LFDT grouped according to \code{groups}}

\item{alpha}{The base significance level of the test.}

\item{groups}{The names of grouping columns in \code{data}}

\item{bonferroni_correct}{Should we Bonferroni correct the significance level based the number of MS comparisons. Default is \code{TRUE}}

\item{conservative}{Should the Bonferroni correction be done conservatively that assumes that some of the MS comparisons are dependent? Default is \code{TRUE}}
}
\value{
A data table with grouping as the first column, LM test statistic as the second column, p-value as the third column, and judgement as the last column
}
\description{
Checks normality of error terms by using Shapiro Wilk's test on each groups' residuals
}
\examples{
print(1)
}
