% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_precision_against_concentration.R
\name{plot_precision_against_concentration}
\alias{plot_precision_against_concentration}
\title{plot precision against concentration}
\usage{
plot_precision_against_concentration(
  data,
  units = "mmol/L",
  save = FALSE,
  save_path = getwd(),
  save_name = "Auto",
  save_device = "pdf",
  size = 8
)
}
\arguments{
\item{data}{A data table with two ID-columns, namely \code{Comparison} and \code{SampleID}}

\item{units}{The units of the analyte considered. These are included in the x-axis name of the plots}

\item{save}{Save plot to disk. Default is \code{FALSE}. If \code{TRUE}, you may specify saving location, file and other with the next four parameters}

\item{save_path}{Saving location given by a string. Default is the working directory}

\item{save_name}{Name of the plot used to save. If \code{save_name = 'Auto'}, a generic name is used based on \code{data}}

\item{save_device}{What file do you want to save the plot as. Valid inputs are 'pdf', 'png', 'jpeg', and more. Using 'png' will produce plots with retina \code{dpi}}

\item{size}{What should the size of the saved plot be? The number specified is the height of the plot. The width of the plot is calculated based on this and the golden ratio}
}
\value{
A plot (\code{ggplot2} object) based on data. Use \code{lapply()} to make multiple plots at once. Then combine the plots using the \code{egg} package.
}
\description{
Plot nice-looking plots of CS-wise precision against concentration. This function is a convenience function that makes plotting much more efficient than doing it manually.
}
\details{
If the \code{Measure} column in \code{data} is lambda, grouping colors will not be included because no groups are available. If more than one comparison is considered, \code{data} will go through the function \code{facet_wrap()}
}
\examples{
plot_precision_against_concentration(data = )
}
