% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residual_histograms.R
\name{plot_residual_histograms}
\alias{plot_residual_histograms}
\title{Plotting of histograms of residuals to visualize the normality assumption}
\usage{
plot_residual_histograms(
  data,
  raw = FALSE,
  MOR = TRUE,
  groups = "Comparison",
  method = "SNRR",
  units = "mmol/L",
  studentize = TRUE,
  include_curve = TRUE,
  save = FALSE,
  save_path = getwd(),
  save_name = "Auto",
  save_device = "pdf",
  size = 8
)
}
\arguments{
\item{data}{A data table or data frame with data generated from \code{residual_histogram_data}, where Comparison, residuals, standardized_residuals and more are included. If raw data of type LFDT is the input, it will be converted to residual histo gram data. To allow such transformation set \code{raw} to \code{TRUE}, and use \code{MOR} and \code{groups} argument to control the transformation}

\item{raw}{Is \code{data} raw, or is it on the appropriate form described in the \code{data} argument}

\item{MOR}{Is your data table containing mean of replicates? Only need to specify if \code{raw = TRUE}}

\item{groups}{What are the grouping column names? Only need to specify if \code{raw = TRUE}}

\item{method}{What method should be used to calculate the optimal bandwidths}

\item{units}{The units of the analyte considered. These are included in the x-axis and y-axis names of the plots}

\item{studentize}{Should we studentize the raw residuals. Default is \code{TRUE}. If \code{TRUE}, you may choose if a N(0,1) density curve should be drawn. This may be triggered by setting \code{include_curve = TRUE}}

\item{include_curve}{Should a density curve of a N(0,1) be displayed over the histogram to visualize the assumed normality. Only relevant if \code{studentize = TRUE}}

\item{save}{Save plot to disk. Default is \code{FALSE}. If \code{TRUE}, you may specify saving location, file and other with the next four parameters}

\item{save_path}{Saving location given by a string. Default is the working directory}

\item{save_name}{Name of the plot used to save. If \code{save_name = 'Auto'}, a generic name is used based on \code{data}}

\item{save_device}{What file do you want to save the plot as. Valid inputs are 'pdf', 'png', 'jpeg', and more. Using 'png' will produce plots with retina \code{dpi}}

\item{size}{What should the size of the saved plot be? The number specified is the height of the plot. The width of the plot is calculated based on this and the golden ratio}
}
\value{
A plot (\code{ggplot2} object) based on data
}
\description{
Plotting of histograms of residuals to visualize the normality assumption
}
\examples{
print(1)
}
