% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_residual_vs_fitted.R
\name{plot_residual_vs_fitted}
\alias{plot_residual_vs_fitted}
\title{Plot residual vs. fitted plots for your data}
\usage{
plot_residual_vs_fitted(
  data,
  raw = TRUE,
  MOR = TRUE,
  groups = "Comparison",
  units = "mmol/L",
  studentize = TRUE,
  include_curve = TRUE,
  save = FALSE,
  save_path = getwd(),
  save_name = "Auto",
  save_device = "pdf",
  size = 8
)
}
\arguments{
\item{data}{A data table or data frame grouped by Comparison where for each comparison, residuals and fitted values are included. If you use a raw data set of type LFDT you must specify \code{MOR}, \code{raw} and \code{groups} correctly. See \code{residual_vs_fitted()} for more information}

\item{raw}{Is \code{data} raw, or is it on the appropriate form described in the \code{data} argument}

\item{MOR}{Is your data table containing mean of replicates? Only need to specify if \code{raw = TRUE}}

\item{groups}{What are the grouping column names? Only need to specify if \code{raw = TRUE}}

\item{units}{The units of the analyte considered. These are included in the x-axis and y-axis names of the plots}

\item{studentize}{Should we studentize the residuals. Default is \code{TRUE}}

\item{include_curve}{Should a smoothing curve be included? Default is \code{TRUE}}

\item{save}{Save plot to disk. Default is \code{FALSE}. If \code{TRUE}, you may specify saving location, file and other with the next four parameters}

\item{save_path}{Saving location given by a string. Default is the working directory}

\item{save_name}{Name of the plot used to save. If \code{save_name = 'Auto'}, a generic name is used based on \code{data}}

\item{save_device}{What file do you want to save the plot as. Valid inputs are 'pdf', 'png', 'jpeg', and more. Using 'png' will produce plots with retina \code{dpi}}

\item{size}{What should the size of the saved plot be? The number specified is the height of the plot. The width of the plot is calculated based on this and the golden ratio}
}
\value{
A plot (\code{ggplot2} object) based on data.
}
\description{
Plot residual vs. fitted plots for your data
}
\examples{
print(1)
}
