% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_of_replicates.R
\name{precision_of_replicates}
\alias{precision_of_replicates}
\title{Estimate precision estimates for each clinical sample over groups}
\usage{
precision_of_replicates(
  data,
  measure = c("standard deviation", "coefficient of variation", "lambda")
)
}
\arguments{
\item{data}{A data table or data frame with LFDT format. Must at least contain MP_A, MP_B and something that resembles ID columns}

\item{measure}{Which precision measure should be used to calculate precision of clinical samples against concentration range? Valid inputs are 'standard deviation' (alternatively, 1 or SD), 'coefficient of variation' (alternatively, 2 or CV) or 'lambda' (alternatively, 3 or L)}
}
\value{
Returns a grouped data table with concentration and precision estimates. If \code{measure = 'lambda'}, the MS column will not be included
}
\description{
Estimate precision estimates for each clinical sample over groups
}
\examples{
precision_of_replicates(MS_wise(sampled_cs_measurements), measure = 'cv')
}
