% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_commutability_evaluation_results.R
\name{present_commutability_evaluation_results}
\alias{present_commutability_evaluation_results}
\title{Present every evaluated materials' commutability for every unique measurement system}
\usage{
present_commutability_evaluation_results(
  data,
  evaluated_materials,
  LFDT = FALSE,
  R = 3,
  level = 0.99,
  method = "DGF",
  success = "Commutable",
  failure = "Non-commutable"
)
}
\arguments{
\item{data}{A data frame or data table with format LFDT or WFDT enclosed with all replicated measurements. This is needed for computation}

\item{evaluated_materials}{A data frame or data table with format LFDT or WFDT enclosed with all replicated measurements for evaluated materials such as EQAMs or CRMs. This is needed for computation}

\item{LFDT}{Is data and evaluated_materials of format LDFT or WFDT?}

\item{R}{An integer signifying the maximum number of replicated measurements performed on each evaluated material. This will shrink the width of the prediction interval and band because using the mean of replicates decreases uncertainty}

\item{level}{A numeric value that captures the overall confidence level of the estimated prediction intervals. This is assumed to already be Bonferroni-corrected. The recommended base level is 0.99}

\item{method}{Which regression estimator should be used to calculate prediction intervals. Write "DFG" (Default) for Deming regression with Gillard and Fuller method. Write "DCE" for CLSI method}

\item{success}{In the summary table, what should we denote those pairs of evaluated materials and measurement systems where the evaluated materials are within prediction interval limits}

\item{failure}{In the summary table, what should we denote those pairs of evaluated materials and measurement systems where the evaluated materials are outside prediction interval limits}
}
\value{
A data table with 1 + n columns and m rows. Here n is the number of unique measurement systems considered and m is the number of unique evaluated materials
}
\description{
Present every evaluated materials' commutability for every unique measurement system
}
\examples{
present_commutability_evaluation_results(data = sampled_cs_measurements, evaluated_materials = sampled_eqam_measurements, method = "DCE")
}
