% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_k_values_bs.R
\name{present_k_values_bs}
\alias{present_k_values_bs}
\title{Present calculated k values with bootstrapped confidence interval for all unique measurement system comparisons}
\usage{
present_k_values_bs(
  data,
  groups = "Comparison",
  LFDT = FALSE,
  decimals = 6,
  n = 100,
  level = 0.95,
  seed_id = 1,
  two_sided = FALSE
)
}
\arguments{
\item{data}{A data frame or data table with format LFDT or WFDT (default) enclosed with all replicated measurements}

\item{groups}{Character vector - The names of the grouping columns of the data. Default is 'Comparison'}

\item{LFDT}{Logical - Is data of format LDFT or WFDT (default)?}

\item{decimals}{Integer - how many decimals should be used in the output data table. Default is 6}

\item{n}{The number of bootstrap resamples of k values for each group. Note that choosing a large number may result in a significant computation time. n > 500 is not recommended!}

\item{level}{A numeric value that captures the overall confidence level of bootstrapped confidence intervals}

\item{seed_id}{Integer - Set a unique seed for reproducibility, or generate a random integer based on a distribution}

\item{two_sided}{Logical - should two-sided confidence intervals be calculated. Default is no because of the practical use of k}
}
\value{
A data table with n rows and 2 columns, where n is the number of unique measurement system combinations in data
}
\description{
Present calculated k values with bootstrapped confidence interval for all unique measurement system comparisons
}
\examples{
present_k_values_bs(data = sampled_cs_measurements, n = 10, level = 0.99, seed_id = 5, two_sided = TRUE)
}
