% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/present_MS_precision_estimates_bs.R
\name{present_MS_precision_estimates_bs}
\alias{present_MS_precision_estimates_bs}
\title{Present precision estimates with bootstrapped confidence interval for all unique measurement systems}
\usage{
present_MS_precision_estimates_bs(
  data,
  groups = "Comparison",
  LFDT = FALSE,
  decimals = 6,
  n = 100,
  level = 0.95,
  seed_id = 1
)
}
\arguments{
\item{data}{A data frame or data table with format LFDT or WFDT enclosed with all replicated measurements}

\item{groups}{Character vector - The names of the grouping columns of the data. Default is 'Comparison'}

\item{LFDT}{Logical - Is data of format LDFT or WFDT (default)?}

\item{decimals}{Integer - how many decimals should be used in the output data table. Default is 6}

\item{n}{The number of bootstrap resamples of MS precision values. Note that choosing a large number may result in a significant computation time}

\item{level}{A numeric value that captures the overall confidence level of bootstrapped confidence intervals}

\item{seed_id}{Integer - Set a unique seed for reproducibility, or generate a random integer based on a distribution}
}
\value{
A data table with n rows and 3 columns, where n is the number of unique measurement systems in data
}
\description{
Present precision estimates with bootstrapped confidence interval for all unique measurement systems
}
\examples{
present_MS_precision_estimates_bs(sampled_cs_measurements, decimals = 3, n = 50, level = 0.99, seed_id = round(runif(1,1,100)))
}
