% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_k.R
\name{resample_k}
\alias{resample_k}
\title{Resample k based on data}
\usage{
resample_k(data, n = 100)
}
\arguments{
\item{data}{A data frame or data table with format LFDT with all replicate measurements enclosed}

\item{n}{The number of bootstrap resamples of k. Note that choosing a large number may result in a significant computation time}
}
\value{
A data table with all resampled k values based on input data
}
\description{
Resample k based on data
}
\examples{
resample_k(data = MS_wise(sampled_cs_measurements)[Comparison=="MP4 - MP5",])
}
