% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_k_over_groups.R
\name{resample_k_over_groups}
\alias{resample_k_over_groups}
\title{Resample k values based on grouped data}
\usage{
resample_k_over_groups(data, n = 100, groups = "Comparison")
}
\arguments{
\item{data}{A grouped data frame or data table with format LFDT}

\item{n}{The number of bootstrap resamples of k for each group. Note that choosing a large number may result in a significant computation time}

\item{groups}{The names of the grouping columns of the data}
}
\value{
A grouped data table listing the n resampled k values for each group
}
\description{
Resample k values based on grouped data
}
\examples{
resample_k_over_groups(data = MS_wise(sampled_cs_measurements), n = 10)
}
