% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_MS_precision.R
\name{resample_MS_precision}
\alias{resample_MS_precision}
\title{Resample MS precision values based on data}
\usage{
resample_MS_precision(data, n = 100)
}
\arguments{
\item{data}{A data frame or data table with format LFDT with all replicate measurements enclosed}

\item{n}{The number of bootstrap resamples of MS precision values. Note that choosing a large number may result in a significant computation time}
}
\value{
A data table with all resampled MS precision values based on input data
}
\description{
Resample MS precision values based on data
}
\examples{
resample_MS_precision(data = MS_wise(sampled_cs_measurements)[Comparison=="MP1 - MP2",], n = 10)
}
