% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_histogram_data.R
\name{residual_histogram_data}
\alias{residual_histogram_data}
\title{Prepare data for constructing histograms of residuals}
\usage{
residual_histogram_data(
  data,
  exclude_fitted = TRUE,
  remove_NA = TRUE,
  method = c("FDR", "SNRR", "SF", "RR", "DF")
)
}
\arguments{
\item{data}{A data table or data frame that is an output form \code{residuals_vs_fitted()}}

\item{exclude_fitted}{Should the fitted values be discarded? Default is \code{TRUE}}

\item{remove_NA}{Should NA-values be removed from data. Default is \code{TRUE}}

\item{method}{Which method should be used to obtain the optimal number of bins, that is, the width of the bins. Valid inputs are \code{FDR}, \code{SNRR}, \code{SF}, \code{RR} and \code{DF}. See details for more information regarding the different approaches}
}
\value{
A data table that have new columns relevant for \code{plot_residual_histograms()}
}
\description{
Prepare data for constructing histograms of residuals
}
\examples{
print(1)
}
