% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residual_vs_fitted.R
\name{residual_vs_fitted}
\alias{residual_vs_fitted}
\title{Summarize to residuals vs fitted for each group}
\usage{
residual_vs_fitted(data, groups = "Comparison", MOR = TRUE)
}
\arguments{
\item{data}{A data table or data frame with LFDT format}

\item{groups}{Column name(s) of grouping columns. Default is Comparison}

\item{MOR}{Should mean of replicates be applied to data before residuals and fitted values are calculated? Default is \code{TRUE}}
}
\value{
A grouped data table according to \code{groups}, where in each group, residuals and fitted values are listed
}
\description{
Summarize to residuals vs fitted for each group
}
\examples{
residual_vs_fitted(MS_wise(sampled_cs_measurements))
}
