% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documentation_sampled_cs_measurements.R
\docType{data}
\name{sampled_cs_measurements}
\alias{sampled_cs_measurements}
\title{Measurement results for 25 different clinical samples}
\format{
A data frame with 75 rows and 7 variables:
\describe{
  \item{SampleID}{The ID of the specific clinical sample}
  \item{ReplicateID}{Based on a specific clinical sample, what is the replicate measurement ID}
  \item{MP1}{Measurement results using measurement system number 1}
  \item{MP2}{Measurement results using measurement system number 2}
  \item{MP3}{Measurement results using measurement system number 3}
  \item{MP4}{Measurement results using measurement system number 4}
  \item{MP5}{Measurement results using measurement system number 5}
}
}
\source{
\url{https://www.bipm.org/en/home}
}
\usage{
sampled_cs_measurements
}
\description{
A data set containing the measurements based on five measurements systems done on 25 clinical samples measured in triplicate.
For definitions, see the link below
}
\keyword{datasets}
