% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwinsp.R
\name{sdwinsp}
\alias{sdwinsp}
\title{Simulation of data with identical ns profiles}
\usage{
sdwinsp(n = "r", R = "r", cv_vec = "r", ci = "r", parameter_row = c(1))
}
\arguments{
\item{n}{Integer or 'r' - number of unique samples}

\item{R}{Integer or 'r' - number of unique samples}

\item{cv_vec}{vector with two elements or 'r' - MS CVs in decimal}

\item{ci}{vector with two elements or 'r' - Concentration interval}

\item{parameter_row}{A data frame or data table with only 1 row. Should contain a subset of simulation parameters}
}
\value{
A data table with 4 columns, two of which are measurement results from two MSs in comparison. In addition, n * R rows
}
\description{
Simulation of data with identical selectivity selectivity profiles. To be used in simulations. Setting either simulation parameters to 'r' will invoke random generation of these numbers. Note that specifying parameters both outside and inside of parameter_row will make those parameters outside being overridden.
}
\details{
Combine with \code{lapply()} or \code{sapply()} to simulate more than one data table. For parallel computing use \code{mc_sdwinsp()}. This function should only be used by advanced users. It is not user friendly for those not familiar with R
}
\examples{
sdwinsp()
}
