% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwinsp2.R
\name{sdwinsp2}
\alias{sdwinsp2}
\title{Simulation of heteroscedastic data with identical non-selectivity profiles}
\usage{
sdwinsp2(
  n = "r",
  R = "r",
  cv_vec = "r",
  ci = "r",
  xi0 = 1,
  xi = 2,
  parameter_row = c(1)
)
}
\arguments{
\item{n}{Integer or 'r' - number of unique samples}

\item{R}{Integer or 'r' - number of unique replicated measurements of each sample}

\item{cv_vec}{vector with two elements or 'r' - MS CVs in decimal}

\item{ci}{vector with two elements or 'r' - Concentration interval}

\item{xi0}{number that corresponds to starting SD respect to base SD}

\item{xi}{number that corresponds to ending SD respect to starting SD}

\item{parameter_row}{A data frame or data table with only 1 row. Should contain a subset of simulation parameters. Possible simulation parameters are n, R, cvx, cvy, ci_lwr and ci_upr, xi0, xi}
}
\value{
A data table with 4 columns, two of which are measurement results from two MSs in comparison. In addition, n * R rows
}
\description{
Simulation of data with identical selectivity profiles, but with heteroscedasticity. To be used in simulations. Setting either simulation parameters to 'r' will invoke random generation of these numbers. Note that specifying parameters both outside and inside of parameter_row will make those parameters outside being overridden.
}
\details{
Combine with \code{lapply()} or \code{sapply()} to simulate more than one data table. For parallel computing use \code{mc_sdwinsp2()}. This function should only be used by advanced users. It is not user friendly for those not familiar with R
}
\examples{
sdwinsp2()
}
