% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_k.R
\name{simulate_k}
\alias{simulate_k}
\title{Simulate k values based on simulation function}
\usage{
simulate_k(
  m = 100,
  method = "sdwinsp",
  parameters,
  to_dt = TRUE,
  parallel = TRUE
)
}
\arguments{
\item{m}{The number of simulated k values for each row of \code{parameters}}

\item{method}{The name of the function of type mc_* that you want to simulate k with. Allowed methods are \code{sdwinsp} (default), \code{sdwinsp2}, \code{sdwdnsp1}, \code{sdwdnsp2}, or 1, 2, 3, 4 respectively}

\item{parameters}{A data table or data frame where each row include a set of parameters. To generate a data table of unique combinations for a set of vectors of parameters it may be used to employ \code{CJ()} from the \code{data.table} package}

\item{to_dt}{If TRUE resulting k values are gathered into an data table}
}
\value{
A list or data table consisting of simulated k values based on generic parameters such as n, R, etc., and method dependent parameters. If \code{to_dt} is TRUE, the parameters found in \code{parameters} will be attached to the output data table
}
\description{
A wrapper function for simulation of k values based on using either mc_* functions to simulate data. Note that all parameters not found in \code{parameters} and is relevant for the given \code{method} will be randomly generated or fixed. Special method dependent parameters such as xi, mmax and q will be fixed if not given in \code{parameters}
}
\examples{
simulate_k(m = 1e3, method = 3, parameters = data.frame(n = c(20,30), R = c(3, 3), p = c(0.05, 0.05), mmax = c(5,5)), to_dt = TRUE)
}
