% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/studentize.R
\name{studentize}
\alias{studentize}
\title{Studentize data using the standard technique}
\usage{
studentize(data)
}
\arguments{
\item{data}{A numeric vector that we want to studentize}
}
\value{
A numeric vector that has studentized values
}
\description{
Studentize data using the standard technique
}
\examples{
studentize(rnorm(100,1,1))
}
