% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_sdwinsp2.R
\name{mc_sdwinsp2}
\alias{mc_sdwinsp2}
\title{Simulate several data sets with identical non-selectivity profiles using parallel computation}
\usage{
mc_sdwinsp2(
  parameter_row = c(1),
  m = 100,
  n = "r",
  R = "r",
  cv_vec = "r",
  ci = "r",
  xi0 = 1,
  xi = 2
)
}
\arguments{
\item{parameter_row}{A data frame or data table with only 1 row. Should contain a subset of simulation parameters. Possible simulation parameters are n, R, cvx, cvy, ci_lwr and ci_upr, xi0, xi}

\item{m}{Integer - The number of data sets to be simulated}

\item{n}{Integer or 'r' - number of unique samples}

\item{R}{Integer or 'r' - Number of unique replicated measurements on each sample}

\item{cv_vec}{Vector with two elements or 'r' - MS CVs in decimal}

\item{ci}{Vector with two elements or 'r' - Concentration interval}

\item{xi0}{A number signifying the relative starting value to the base. Default is 1}

\item{xi}{A number signifying the relative ending value relative to the base. Default is 2}
}
\value{
A list containing m simulated data tables with 4 columns, two of which are measured results from two MSs in comparison. In addition, n * R rows.
}
\description{
Simulate several data sets with identical non-selectivity profiles using parallel computation
}
\details{
this function is recommended to use when simulating more than 500 data sets, or else it will be faster to just use standard sapply
}
\examples{
mc_sdwinsp2(m = 50, xi0 = 1, xi = 3)
}
