% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS_wise.R
\name{MS_wise}
\alias{MS_wise}
\title{Converting standard wide format MS comparison data (WFDT) to standard long format (LFDT)}
\usage{
MS_wise(data, force_IDs = NULL)
}
\arguments{
\item{data}{A data frame or data table with m + n columns where m (usually two) of them are ID columns and n of them is numeric columns consisting of measurement results for each measurement system}

\item{force_IDs}{If you want to type in the ID column names manually, you may specify them here. This is only recommended if the column ID names are far from the standard names. For example, if the Sample ID column is named 'WDXQXX' and Replicate ID column 'RGDHEC', or is in a language very different from Scandinavian languages and English}
}
\value{
A data table of format LFDT, with m + 3 columns where the latter three columns are MS comparison ID column (new) and resulting numeric columns in this comparison
}
\description{
Converting standard wide format MS comparison data (WFDT) to standard long format (LFDT)
}
\details{
If you specify \code{force_IDs} to your ID column names, the output will keep these names. It is then recommended to run \code{fix_column_names_2(data = output, n_ids = 3, LFDT = TRUE, MOR = FALSE, slack = 4)} to make the output into the standard form so that further functions may be used
}
\examples{
MS_wise(sampled_eqam_measurements)
}
