% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdwinsp2.R
\name{sdwinsp2}
\alias{sdwinsp2}
\title{Simulation of data with identical ns profiles}
\usage{
sdwinsp2(
  n = "r",
  R = "r",
  cv_vec = "r",
  ci = "r",
  xi0 = 1,
  xi = 2,
  parameter_row = c(1)
)
}
\arguments{
\item{n}{Integer or 'r' - number of unique samples}

\item{R}{Integer or 'r' - number of unique samples}

\item{cv_vec}{vector with two elements or 'r' - MS CVs in decimal}

\item{ci}{vector with two elements or 'r' - Concentration interval}

\item{xi0}{number that corresponds to starting SD respect to base SD}

\item{xi}{number that corresponds to ending SD respect to starting SD}

\item{parameter_row}{A data frame or data table with only 1 row. Should contain a subset of simulation parameters}
}
\value{
A data table with 4 columns, two of which are measurement results from two MSs in comparison. In addition, n * R rows
}
\description{
Simulation of data without non-selectivity but with heteroscedasticity. To be used in simulations
}
\details{
Combine with lapply to simulate more than one data table. This function should only be used by advanced users. It is not user friendly for those not familiar with R
}
\examples{
sdwinsp2()
}
